added main text, and moved message to exc. map
