+++ src/main/java/app/exceptions/Message.java
@@ -1,3 +1,3 @@
-package app.dtos;
+package app.exceptions;
 
 public record Message(int status, String message) {}
